Autopredict addition
So this contains barebones autoprediction. But for some reason only the first
link is clickable. I'm not sure why.
In newresults.tpl, when you enter into a textbox, it goes into that function and sends a query
back to our run.py which has an autopredict module that is loaded already (which has been machine trained) over
some million words. So that parses the last two words of the query and predicts the next word and suggests it